+++ packages/xchain-thorchain-query/__mocks__/midgard-api.ts
@@ -1,6 +1,7 @@
 import mock from './axios-adapter'
 
 export default {
+  reset: mock.reset,
   restore: mock.restore,
   init: () => {
     //Mock midgard pools
@@ -9,28 +10,34 @@ export default {
       return [200, resp]
     })
     //Mock thorchain/inbound_addresses
-    mock.onGet(/\/thorchain\/inbound_addresses/).reply(function () {
+    mock.onGet(/\/v2\/thorchain\/inbound_addresses/).reply(function () {
       const resp = require(`./responses/midgard/inbound_addresses.json`)
       return [200, resp]
     })
     //Mock thorchain/mimir
-    mock.onGet(/\/thorchain\/mimir/).reply(function () {
+    mock.onGet(/\/v2\/thorchain\/mimir/).reply(function () {
       const resp = require(`./responses/mimir/mimirConstants.json`)
       return [200, resp]
     })
     //Mock thorchain/mimir
-    mock.onGet(/\/thorchain\/constants/).reply(function () {
+    mock.onGet(/\/v2\/thorchain\/constants/).reply(function () {
       const resp = require(`./responses/thorchain/thorchainConstants.json`)
       return [200, resp]
     })
     //Mock thorchain/mimir
-    mock.onGet(/\/thorchain\/queue/).reply(function () {
+    mock.onGet(/\/v2\/thorchain\/queue/).reply(function () {
       const resp = require(`./responses/thorchain/outboundQueue.json`)
       return [200, resp]
     })
     //Mock midgard health
     mock.onGet(/\/v2\/health/).reply(function () {
       const resp = require(`./responses/midgard/health.json`)
+      return [200, resp]
+    })
+    //Mock midgard actions
+    mock.onGet(/\/v2\/actions?/).reply(function () {
+      const resp = require(`./responses/midgard/health.json`)
+      console.log(`actions`)
       return [200, resp]
     })
   },