+++ packages/xchain-thorchain-amm/CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.0.1.0-beta4 (2022-09-15)
+
+- moved examples into different directory
+- import new version of xchain-thorchain-query
+
 # v0.0.1.0-beta3 (2022-09-06)
 
 - import new version of xchain-thorchain-query